service1/requirements.txt
Flask==1.1.1
markupsafe==2.0.1
itsdangerous==2.0.1
Jinja2==3.0.3
flask_sqlalchemy
pymysql
requests
werkzeug==2.0.3 